Add a new Assertable mechanism.
Provides a way to considerably reduce boilerplate in HTTP endpoint handlers,
by using an assert/exeption model. This is based on a (mis?)use of
panic/defer, and works hand-in-hand with the existing Wrapper/Sentry system
(and specifically, will not be useful outside that wrapper system.)

Also adds logging to the Wrapper.